backend core #33: move old template files to old/ folder
